Product groups handbook: Air guitar
We want to think further than one sprint when designing some features.

For example, we started drafting Fleet’s app library last sprint but we don’t plan on estimating it and building it in the current sprint. So, it won’t ship in 6 weeks. Today, we say that everything on the drafting board ships in 6 weeks.

An air guitar can be used to designing features further than one sprint out.