Warn on CVE description language tags from NVD feed that we haven't seen before
We may need more effort to alert on this in a place where we can see/action it, but for that alerting we can just catch warn (or warn-and-above) logs, as I just ran the artifact build command and didn't get any warnings.

Confirmed that this works by starting without the "es" case (since I wasn't sure which language tags other than "en" and "en-US" we were seeing) and seeing a bunch of CVEs from 2004 with "es" language tags reported as warnings. Also confirmed that language tags were correctly set in the resulting feed (caught a bug locally due to Golang having an implied break at the end of each switch case).