Implement targeted certificate patching in fleetd
**NOT TO BE MERGED**

This is to help out a customer who has deployed fleetd agents with a single certificate that is expiring. We look for that certificate and replace it with the default cert bundle. This will be pushed to an orbit release tag that the customer will configure on their Fleet server, so that it only affects their deployment.