consolidate Automatic Enrollment and VPP settings into the MDM settings section in UI (#21272)
relates to #21184

This consolidates the Automatic enrollment and VPP settings into the MDM
settings section on the integrations page. It also includes copy changes
and changes in how we handle loading states and error states.

> NOTE: this included moving around of a lot of the components that were
separate sections before. The organizational pattern has a `MdmSettings`
directory as the parent. In this directory there is a `component`
directory with all the section components. as well as the pages related
to those sections e.g.:

```
MdmSettings
|_AppleMdmPage
|_AppleAutomaticEnrollmentPage
|_VppSetupPage
|_WindowsAutomaticEnrollmentPage
|_components
  |_MdmSettingsSection
  |_AutomaticEnrollmentSection
  |_VppSection
...
```

> ANOTHER NOTE: Please pay special attention to the rendering of the
content on the `MdmSettings.tsx` file. The desired behaviour is that for
**ANY** error we want to show the error message and we want to show the
loading message when all requests are in progress. I tried to make it
clear what was needed to rendering but a second pair of eyes here would
be useful

**auto enrollment and vpp settings in mdm settings section**


![image](https://github.com/user-attachments/assets/1d8cbc27-d79f-499c-b907-765f1bd69672)

**idp, eula, and end user migration also in mdm settings section**


![image](https://github.com/user-attachments/assets/35e6c71e-79ca-4673-af0d-61b3fff51bf1)


- [x] Changes file added for user-visible changes in `changes/`,
`orbit/changes/` or `ee/fleetd-chrome/changes`.
See [Changes
files](https://github.com/fleetdm/fleet/blob/main/docs/Contributing/Committing-Changes.md#changes-files)
for more information.
- [ ] Added/updated tests
- [x] Manual QA for all new/changed functionality