Help contributors find wireframes (#21874)
> Hard to find existing UI elements (and pages) in Figma files. As a
Product Designer, I have to know what the most recent feature, and thus
what issue number, was in which a specific UI element was updated.

Re @randy-fleet's feedback ^

Latest plan was discussed during product design sync. This is instead of
the previous plan: spreadsheet where each row is a page/modal in the UI,
and we link to Figma files as designs get settled.