Use sync.Map for stubbed key-value store to avoid data races in GitOps test (#22292)
This override only happens in testing, so this isn't release-blocking,
but this is the quickest way to clean up a test that will otherwise be
flaky due to data races, at the cost of performance (vs. setting up a
more complex solution with mutexes).

# Checklist for submitter

If some of the following don't apply, delete the relevant line.

<!-- Note that API documentation changes are now addressed by the
product design team. -->

- [x] Added/updated tests
- [x] Manual QA for all new/changed functionality (via manually running
test using the KV store)