Don't reuse GitHub HTTP client to pull MSRC feeds
As of today, MSRC's API will 401 if you pass it a JWT it doesn't expect, and by reusing the GitHub API HTTP client for MSRC pulls we were passing the API an unexpected JWT. Wasn't able to reproduce this locally because I didn't need a GitHub token locally to pull release details, while the token is populated in Actions. Was able to repro both the issue and this fix inside Actions on my fork.

This also updates to call v3.0 of the API directly, which v2.0 has been redirecting to for awhile.

Finally, adds slightly better logging so we know which part of the feed generation process we're in when we're running this in Actions.