Ensure scripts set in no-team.yml can be used in run-script actions for No Team
Also revises the spec check to explain that scripts have to be defined "controls" when used in policies for the same team, with an explicit call-out for no-team.yml since this fix doesn't support pulling scripts from the global file. This is because parsing and script-matching happens early enough that we can't throw an error in the part of the code where we bail when controls are defined in both no-team and default files.

To minimize diff size, we're both passing-by-ref and returning the maps-by-team of scripts and installers, though pass-by-ref would be sufficient on its own. If we're okay with a bigger diff the extra pieces of the return tuple can be dropped in a subsequent revision.