Update render.yaml (#21909)
If a "plan" is not defined, render will default to its "starter"
instance type. For MySQL this means an instance with 0.5 CPU and 512MBs
of RAM. In my testing with only a few hosts enrolled, this resulted in
the service regularly running out of memory and needing to restart.
Bumping this up to a "standard" instance type with 1 CPU and 2 GBs of
RAM has shown to be more stable.

Redis continues to operate without issue on the "starter" instance type
but included defining this in the yaml so it can be updated in the
future if needed.