Switch Nudge deadline from 4am UTC to 8pm UTC (noon Pacific when not on DST)
This only affects macOS hosts below version 14, as we enforce OS updates via DDM for newer versions. We're making a tradeoff of consistent behavior with DDM (which would require trickier fixes) for a smaller patch, while moving the deadline so that it's not happening late on the day before.

Revised the variable name here because we are *not* using the host's local time zone to determine the timestamp here.