Added "a" to "Why a handbook-first strategy"
I wish I could come up with the grammatical reason for this, but It feels like the article is missing. Feel free to disregard if you disagree. Or if you know the grammatical rule please lmk!

Text added on line 34

Link for reference: https://fleetdm.com/handbook/company/why-this-way#why-handbook-first-strategy