Document new without_vulnerability_details option for software versions endpoint (#24246)
#23679

In order to improve performance on the software version endpoint the
option `without_vulnerability_details` has been added. It can be set to
true to decrease the size of the response.

On the free tier setting this option to `false` will have no effect. On
Fleet Premium setting it to `false` or omitting will include verbose
vulnerability details.

---------

Co-authored-by: Rachael Shaw <r@rachael.wtf>