API Docs: Match request format for script edit vs. script add (#24679)
For #24195.

The existing POST /scripts method uses a multipart request, and I figure
that since we use multipart for both create and edit on installers, we
want the same level of consistency on scripts.

If we _don't_ want this, and want a JSON payload instead, we probably
still want to rename `script_contents` to `script` for consistency with
the field name in the create API.