Add note about script execution timeout config (#25189)
relates to #24247 

There's nothing actually wrong in the current docs. We document that to
set `script_execution_timeout` in a YAML file you do:

```yaml
agent_options:
  script_execution_timeout: 600
```

which is correct. It's just that in the UI we hide the top-level
`agent_options` key entirely. This makes sense from a UX standpoint but
it's easy to see how someone could get confused. The _best_ option would
be to have `script_execution_timeout` always returned from the config
endpoint so that it appeared in the right place in the UI and the user
would just have to edit it, but I'd wait for more reports on this before
committing that level of effort.