Add bundle ID based CPE product matching for more JetBrains macOS products
For #22723.

Bundle IDs pulled from installs of the software, with regexes being starts-with matches to include EAPs in matches. Used the products list from CVE-2024-37051 to match up NVD product names (which is why Writerside isn't included here; it doesn't have any published vulns yet).

This fixes vuln detection in e.g. GoLand EAPs when the app name is something other than the product name, similar to what we've done with IntelliJ and PyCharm (but omitting homebrew handling for now).