Document existing orbit auto-uninstall behavior (runs if post-install is supplied and fails, not if install fails)
Haven't done full research yet on why this is, but my guess is that if an install fails we don't know at what stage it failed, so running an uninstall script would potentially just break things further. Putting the trigger on post-install means that it's opt-in (we don't populate post-install scripts by default) and we're uninstalling from a relatively known state (the installer didn't error).

I'll get an issue filed tomorrow to make sure that this remains the desired behavior, or if not then when we change this (e.g. when implementing patching flows), but might as well fix docs in the interim to be clear about what we're doing.