add telemetry for fleet desktop failing to launch because of missing domain (#25752)
For #25689, #19499

We wanted to add some additional telemetry for fleet desktop to know
when it was failing to launch based on the error message from a [recent
bug] https://github.com/fleetdm/fleet/issues/25689. This bug and some
related ones seemed intermittent and we were unable to reproduced so we
are hoping adding this telemetry to give some more analytics on this
issue.