Only use one nginx worker, add a slight wait post-nginx-reload, to properly upgrade-test schema changes that aren't backward-compatible on the "get enroll secret" code path (#26008)
For #26003. Will add more context to that ticket.

This is a robustness fix to the upgrade test, which pulls tags/branches
from Docker, so this does _not_ need to be cherry-picked.

To confirm this fix, run the upgrade test from <4.63.0 to >= 4.63.0 and
see it pass rather than fail.

# Checklist for submitter
- [x] Added/updated automated tests
- [x] A detailed QA plan exists on the associated ticket (if it isn't
there, work with the product group's QA engineer to add it)
- [x] Manual QA for all new/changed functionality