Allow for auto-install for FMAs
Note that this does *not* use the queries in apps.json, as those queries are identical to what we would build programmatically by bundle ID, and this is a smaller change. We'll almost certainly need to revise this for Windows and/or FMAv2, so going with a reliable but closed-ended change for now.