Update docs to reflect revised software add UI, VPP automatic installs
Also switches name-based to bundle-ID-based for install queries since that's what we use for macOS apps, shows how to find the bundle ID (API-only for now), and removes language around these being the queries we use because we don't do a version comparison yet.

Not trying to add screenshots here as the software add UX will change in the next sprint or so.