Split "all hosts" and "specific hosts" handling for Windows MDM profiles to install/remove
The vast majority of the overlap between "all hosts" and "a subset of hosts" for these functions are in the SQL, so if we extract the SQL out we can split the function into one that only operates on a subset of hosts (and optionally a subset of profiles), and one that only operates on all hosts. This reduces the complexity of both functions, which is important given that we just added complexity via batching.

This change winds up with equivalent behavior for both places that call the functions.