Force a user to click through to redeem MFA if they aren't on a browser that started the MFA process
For #26976.

An example of such a browser would be an email link scanner, so this *should* fix cases where link-scanning was redeeming the MFA link before the intended user could get to it. Real users can still click through if they wind up on this page, e.g. if they logged in with a different browser than the one used to open the MFA link.