Add instructions for single sign-on via authentik. (#27079)
Hi Fleet team!

I'm an engineer at authentik and I've been working on a guide to help
our users integrate Fleet with authentik as an identity provider. While
our own documentation is still in progress, I wanted to contribute this
guide to the Fleet documentation to help our users get started with SSO.

This PR expands Fleet's Single Sign-on guide to include a section for
[authentik](https://goauthentik.io) as an identity provider.

Please let me know if I can provide any additional information or make
any changes to this PR.

Thank you for considering this contribution!